Add Pax Levante 50
I recently bought a Pax Levante 50, which according to Pax is the exact same product but geared towards other markets (for some unknown reason).

This fan provides the PIN the same way as described for Svensa. There is no PIN printed on the fan housing. I would not be surprised if new Calima works the same way, but someone with a new Calima would have to confirm that.

So, for the time being, I suggest adding this to the README.md.